+++ sdk/resourcemanager/datafactory/armdatafactory/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release History
 
+## 1.2.0 (2022-06-15)
+### Features Added
+
+- New field `ClientSecret` in struct `RestServiceLinkedServiceTypeProperties`
+- New field `Resource` in struct `RestServiceLinkedServiceTypeProperties`
+- New field `Scope` in struct `RestServiceLinkedServiceTypeProperties`
+- New field `TokenEndpoint` in struct `RestServiceLinkedServiceTypeProperties`
+- New field `ClientID` in struct `RestServiceLinkedServiceTypeProperties`
+
+
 ## 1.1.0 (2022-05-30)
 ### Features Added
 
@@ -14,6 +24,22 @@
 - New struct `GlobalParametersClientGetResponse`
 - New struct `GlobalParametersClientListByFactoryOptions`
 - New struct `GlobalParametersClientListByFactoryResponse`
+
+
+## 1.0.0 (2022-05-17)
+
+The package of `github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/datafactory/armdatafactory` is using our [next generation design principles](https://azure.github.io/azure-sdk/general_introduction.html) since version 1.0.0, which contains breaking changes.
+
+To migrate the existing applications to the latest version, please refer to [Migration Guide](https://aka.ms/azsdk/go/mgmt/migration).
+
+To learn more, please refer to our documentation [Quick Start](https://aka.ms/azsdk/go/mgmt).## 1.2.0 (2022-06-15)
+### Features Added
+
+- New field `ClientSecret` in struct `RestServiceLinkedServiceTypeProperties`
+- New field `Resource` in struct `RestServiceLinkedServiceTypeProperties`
+- New field `Scope` in struct `RestServiceLinkedServiceTypeProperties`
+- New field `TokenEndpoint` in struct `RestServiceLinkedServiceTypeProperties`
+- New field `ClientID` in struct `RestServiceLinkedServiceTypeProperties`
 
 
 ## 1.0.0 (2022-05-17)